Added all the fixtures to the root conftest. Updated the imports.
